fix(price-feed): combined tests to see if ci will pass
